add reset_sampler method
