+++ trieste/models/interfaces.py
@@ -15,14 +15,14 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Callable
 
 import gpflow
 import tensorflow as tf
+from gpflow.sampler import BatchReparameterizationSampler
 
 from ..data import Dataset
 from ..types import TensorType
-from ..utils import DEFAULTS
+from .sampler import ReparametrizationSampler, TrajectorySampler
 
 
 class ProbabilisticModel(ABC):
@@ -266,102 +266,31 @@ class ModelStack(TrainableProbabilisticModel):
 
     def reparam_sampler(self, num_samples: int) -> ReparametrizationSampler:
         """
-        Return a reparametrization sampler providing `num_samples` samples across
-        all the models in the model stack.
+        Return a reparameterization sampler providing `num_samples` samples across
+        all the models in the model stack. This is currently only implemented for
+        stacks made from models that have a :class:`BatchReparameterizationSampler`
+        as their reparameterization sampler.
 
         :param num_samples: The desired number of samples.
-        :return: The reparametrization sampler.
+        :return: The reparameterization sampler.
         :raise NotImplementedError: If the models in the stack do not share the
             same :meth:`reparam_sampler`.
         """
 
-        sampler_types = [type(model.reparam_sampler(1)) for model in self._models]
-        unique_sampler_types = set(sampler_types)
-        if len(unique_sampler_types) > 1:
+        samplers = [model.reparam_sampler(num_samples) for model in self._models]
+
+        if all(isinstance(sampler, BatchReparameterizationSampler) for sampler in samplers):
+            return BatchReparameterizationSampler(
+                num_samples, self
+            )  # return the shared sampler rebuilt for the whole model stack
+        else:
             raise NotImplementedError(
                 f"""
                 Reparameterization sampling is only currently supported for model
-                stacks built from models that use the same reparameterization sampler,
-                however, received samplers of types {unique_sampler_types}.
+                stacks built from models that use the BatchReparameterizationSampler
+                however, received {samplers}.
                 """
             )
-        else:  # return the shared sampler rebuilt for the whole model stack
-            shared_sampler = sampler_types[0]
-            return shared_sampler(num_samples, self)
-
-
-class ReparametrizationSampler(ABC):
-    r"""
-    This sampler employs the *reparameterization trick* to draw samples from a
-    :class:`ProbabilisticModel`\ 's predictive distribution  across a discrete set of
-    points.
-    """
-
-    def __init__(self, sample_size: int, model: ProbabilisticModel):
-        """
-        :param sample_size: The desired number of samples.
-        :param model: The model to sample from.
-        :raise ValueError (or InvalidArgumentError): If ``sample_size`` is not positive.
-        """
-        tf.debugging.assert_positive(sample_size)
-
-        self._sample_size = sample_size
-        self._model = model
-        self._initialized = tf.Variable(False)
-
-    def __repr__(self) -> str:
-        """"""
-        return f"{self.__class__.__name__}({self._sample_size!r}, {self._model!r})"
-
-    @abstractmethod
-    def sample(self, at: TensorType, *, jitter: float = DEFAULTS.JITTER) -> TensorType:
-        """
-        :param at: Input points that define the sampler.
-        :param jitter: The size of the jitter to use when stabilising the Cholesky
-            decomposition of the covariance matrix.
-        :return: Samples.
-        """
-
-
-TrajectoryFunction = Callable[[TensorType], TensorType]
-"""
-Type alias for trajectory functions.
-
-An :const:`TrajectoryFunction` evaluates a particular sample at a set of `N` query
-points (each of dimension `D`) i.e. takes input of shape `[N, D]` and returns
-shape `[N, 1]`.
-
-A key property of these trajectory functions is that the same sample draw is evaluated
-for all queries. This property is known as consistency.
-"""
-
-
-class TrajectorySampler(ABC):
-    r"""
-    This class builds functions that approximate a trajectory sampled from an
-    underlying :class:`ProbabilisticModel`.
-
-    Unlike the :class:`ReparametrizationSampler`, a :class:`TrajectorySampler` provides
-    consistent samples (i.e ensuring that the same sample draw is used for all evaluations
-    of a particular trajectory function).
-    """
-
-    def __init__(self, model: ProbabilisticModel):
-        """
-        :param model: The model to sample from.
-        """
-        self._model = model
-
-    def __repr__(self) -> str:
-        """"""
-        return f"{self.__class__.__name__}({self._model!r})"
-
-    @abstractmethod
-    def get_trajectory(self) -> TrajectoryFunction:
-        """
-        :return: A trajectory function representing an approximate trajectory from the
-            model, taking an input of shape `[N, D]` and returning shape `[N, 1]`
-        """
 
 
 class FastUpdateModel(ABC):
